requirements.txt
sounddevice==0.4.6
numpy==1.24.3
customtkinter==5.2.2
pillow==10.2.0
soundfile==0.12.1
